MAINT: server – clean-up compat and mocking code
